\name{catIrt-package}
\alias{catIrt-package}
\docType{package}
\title{Simulate IRT-Based Computerized Adaptive Tests (CATs)}
\description{
\code{catIrt} provides methods for simulating Computerized Adaptive Tests,
including response simulation, item selection methods, ability estimation methods,
and test termination methods.  Unique in \code{catIrt} is support for the graded
response model (and, soon, other polytomous models) along with expanding support
for classification CAT (including an implementation of the SPRT, GLR, and CI methods).
}

}
\author{
Maintainer: Steven W. Nydick \email{swnydick@gmail.com}
}
