\name{catIrt-package}
\alias{catIrt-package}
\docType{package}
\title{Simulate Computerized Adaptive Tests (CATs)}
\description{
\code{catIrt} provides functions that allow for simulation of Computerized
Adaptive Tests, including response simulation, item selection methods, ability
estimation methods, and test termination methods.  Unique in \code{catIrt} is
support for the graded response model (and, soon, other polytomous models) along
with expanding support for classification CAT (including an implementation of
the SPRT, GLR, and CI methods).
}

\details{
\tabular{ll}{
Package:  \tab catIrt\cr
Type:     \tab Package\cr
Version:  \tab 0.2-1\cr
Date:     \tab 2012-07-03\cr
License:  \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Maintainer: Steven W. Nydick <nydic001@umn.edu>
}
